done with algorithm, on to application
